+++ h3ron/src/collections.rs
@@ -1,10 +1,9 @@
+use std::borrow::Borrow;
 use std::collections::HashSet;
 use std::iter::FromIterator;
 use std::ops::RangeInclusive;
 
 use serde::{Deserialize, Serialize};
-
-use h3ron_h3_sys::H3Index;
 
 use crate::H3Cell;
 use crate::{compact, Index, H3_MAX_RESOLUTION, H3_MIN_RESOLUTION};
@@ -12,18 +11,18 @@ use crate::{compact, Index, H3_MAX_RESOLUTION, H3_MIN_RESOLUTION};
 const H3_RESOLUTION_RANGE_USIZE: RangeInclusive<usize> =
     (H3_MIN_RESOLUTION as usize)..=(H3_MAX_RESOLUTION as usize);
 
-/// structure to keep compacted h3ron indexes to allow more or less efficient
-/// adding of further indexes
+/// structure to keep compacted h3ron cells to allow more or less efficient
+/// adding of further cells
 #[derive(PartialEq, Serialize, Deserialize, Debug)]
-pub struct H3CompactedVec {
-    /// indexes by their resolution. The index of the array is the resolution for the referenced vec
-    indexes_by_resolution: [Vec<H3Index>; H3_MAX_RESOLUTION as usize + 1],
+pub struct CompactedCellVec {
+    /// cells by their resolution. The index of the array is the resolution for the referenced vec
+    cells_by_resolution: [Vec<H3Cell>; H3_MAX_RESOLUTION as usize + 1],
 }
 
-impl<'a> H3CompactedVec {
-    pub fn new() -> H3CompactedVec {
-        H3CompactedVec {
-            indexes_by_resolution: Default::default(),
+impl<'a> CompactedCellVec {
+    pub fn new() -> CompactedCellVec {
+        CompactedCellVec {
+            cells_by_resolution: Default::default(),
         }
     }
 
@@ -35,10 +34,10 @@ impl<'a> H3CompactedVec {
     pub fn append(&mut self, other: &mut Self, compact: bool) {
         let mut resolutions_touched = Vec::new();
         for resolution in H3_RESOLUTION_RANGE_USIZE {
-            if !other.indexes_by_resolution[resolution].is_empty() {
+            if !other.cells_by_resolution[resolution].is_empty() {
                 resolutions_touched.push(resolution);
-                let mut h3indexes = std::mem::take(&mut other.indexes_by_resolution[resolution]);
-                self.indexes_by_resolution[resolution].append(&mut h3indexes);
+                let mut cells = std::mem::take(&mut other.cells_by_resolution[resolution]);
+                self.cells_by_resolution[resolution].append(&mut cells);
             }
         }
         if compact {
@@ -55,86 +54,78 @@ impl<'a> H3CompactedVec {
         )
     }
 
-    /// append the contents of a vector
+    /// append the contents of a vector. The caller is responsible to ensure that
+    /// the append cells all are at resolution `resolution`.
     ///
-    /// Indexes get moved, see Vec::append
-    pub fn append_to_resolution(
-        &mut self,
-        resolution: u8,
-        h3indexes: &mut Vec<H3Index>,
-        compact: bool,
-    ) {
-        self.indexes_by_resolution[resolution as usize].append(h3indexes);
+    /// Cells get moved, this is the same API as `Vec::append`
+    pub fn append_to_resolution(&mut self, resolution: u8, cells: &mut Vec<H3Cell>, compact: bool) {
+        self.cells_by_resolution[resolution as usize].append(cells);
         if compact {
             self.compact_from_resolution_up(resolution as usize, &[]);
         }
     }
 
     pub fn len(&self) -> usize {
-        self.indexes_by_resolution
+        self.cells_by_resolution
             .iter()
-            .fold(0, |acc, h3indexes| acc + h3indexes.len())
+            .fold(0, |acc, cells| acc + cells.len())
     }
 
     /// length of the vectors for all resolutions. The index of the vec is the resolution
     pub fn len_resolutions(&self) -> Vec<usize> {
-        self.indexes_by_resolution.iter().map(|v| v.len()).collect()
+        self.cells_by_resolution.iter().map(|v| v.len()).collect()
     }
 
     pub fn is_empty(&self) -> bool {
         !self
-            .indexes_by_resolution
+            .cells_by_resolution
             .iter()
-            .any(|h3indexes| !h3indexes.is_empty())
+            .any(|cells| !cells.is_empty())
     }
 
-    /// check if the stack contains the index or any of its parents
+    /// check if the stack contains the cell or any of its parents
     ///
     /// This function is pretty inefficient.
-    pub fn contains(&self, h3index: H3Index) -> bool {
+    pub fn contains(&self, mut cell: H3Cell) -> bool {
         if self.is_empty() {
             return false;
         }
-        let mut index = H3Cell::new(h3index);
-        for r in index.resolution()..=H3_MIN_RESOLUTION {
-            index = match index.get_parent(r) {
+        for r in cell.resolution()..=H3_MIN_RESOLUTION {
+            cell = match cell.get_parent(r) {
                 Ok(i) => i,
                 Err(_) => continue,
             };
-            if self.indexes_by_resolution[r as usize].contains(&index.h3index()) {
+            if self.cells_by_resolution[r as usize].contains(&cell) {
                 return true;
             }
         }
         false
     }
 
-    /// add a single h3 index
+    /// add a single h3 cell
     ///
     /// will trigger a re-compacting when `compact` is set
-    pub fn add_index(&mut self, h3_index: H3Index, compact: bool) {
-        let resolution = H3Cell::new(h3_index).resolution();
-        self.add_index_to_resolution(h3_index, resolution, compact);
-    }
-
-    /// add a single h3ron index
-    ///
-    /// the `resolution` parameter must match the resolution of the index. This method
-    ///  only exists to skip the resolution check of `add_index`.
-    pub fn add_index_to_resolution(&mut self, h3_index: H3Index, resolution: u8, compact: bool) {
-        self.indexes_by_resolution[resolution as usize].push(h3_index);
+    #[inline]
+    pub fn add_cell(&mut self, cell: H3Cell, compact: bool) {
+        let res = cell.resolution() as usize;
+        self.cells_by_resolution[res].push(cell);
         if compact {
-            self.compact_from_resolution_up(resolution as usize, &[]);
+            self.compact_from_resolution_up(res, &[]);
         }
     }
 
     ///
     ///
-    pub fn add_indexes(&mut self, h3_indexes: &[H3Index], compact: bool) {
+    pub fn add_cells<I>(&mut self, cells: I, compact: bool)
+    where
+        I: IntoIterator,
+        I::Item: Borrow<H3Cell> + Index,
+    {
         let mut resolutions_touched = HashSet::new();
-        for h3_index in h3_indexes {
-            let res = H3Cell::new(*h3_index).resolution() as usize;
+        for cell in cells {
+            let res = cell.resolution() as usize;
             resolutions_touched.insert(res);
-            self.indexes_by_resolution[res].push(*h3_index);
+            self.cells_by_resolution[res].push(*(cell.borrow()));
         }
 
         if compact {
@@ -148,45 +139,31 @@ impl<'a> H3CompactedVec {
         }
     }
 
-    pub fn add_indexes_from_iter<T: IntoIterator<Item = H3Index>>(
-        &mut self,
-        iter: T,
-        compact: bool,
-    ) {
-        let mut cv = Self::new();
-        for h3index in iter {
-            self.add_index(h3index, false);
-        }
-        if compact {
-            cv.compact();
-        }
-    }
-
     /// iterate over the compacted (or not, depending on if `compact` was called) contents
-    pub fn iter_compacted_indexes(&self) -> H3CompactedVecCompactedIterator {
-        H3CompactedVecCompactedIterator {
+    pub fn iter_compacted_cells(&self) -> CompactedCellVecCompactedIterator {
+        CompactedCellVecCompactedIterator {
             compacted_vec: self,
             current_resolution: H3_MIN_RESOLUTION as usize,
             current_pos: 0,
         }
     }
 
-    /// get the compacted indexes of the given resolution
+    /// get the compacted cells of the given resolution
     ///
-    /// parent indexes at lower resolutions will not be uncompacted
-    pub fn get_compacted_indexes_at_resolution(&self, resolution: u8) -> &[H3Index] {
-        &self.indexes_by_resolution[resolution as usize]
+    /// parent cells at lower resolutions will not be uncompacted
+    pub fn get_compacted_cells_at_resolution(&self, resolution: u8) -> &[H3Cell] {
+        &self.cells_by_resolution[resolution as usize]
     }
 
-    /// iterate over the uncompacted indexes.
+    /// iterate over the uncompacted cells.
     ///
-    /// indexes at lower resolutions will be decompacted, indexes at higher resolutions will be
+    /// cells at lower resolutions will be decompacted, cells at higher resolutions will be
     /// ignored.
-    pub fn iter_uncompacted_indexes(
+    pub fn iter_uncompacted_cells(
         &self,
         resolution: u8,
-    ) -> H3CompactedVecUncompactedIterator<'_> {
-        H3CompactedVecUncompactedIterator {
+    ) -> CompactedCellVecUncompactedIterator<'_> {
+        CompactedCellVecUncompactedIterator {
             compacted_vec: self,
             current_resolution: H3_MIN_RESOLUTION as usize,
             current_pos: 0,
@@ -195,11 +172,11 @@ impl<'a> H3CompactedVec {
         }
     }
 
-    /// deduplicate the internal h3index vector
+    /// deduplicate the internal cell vectors
     pub fn dedup(&mut self) {
-        self.indexes_by_resolution.iter_mut().for_each(|indexes| {
-            indexes.sort_unstable();
-            indexes.dedup();
+        self.cells_by_resolution.iter_mut().for_each(|cells| {
+            cells.sort_unstable();
+            cells.dedup();
         });
         self.purge_children();
     }
@@ -207,7 +184,7 @@ impl<'a> H3CompactedVec {
     /// the finest resolution contained
     pub fn finest_resolution_contained(&self) -> Option<u8> {
         for resolution in H3_RESOLUTION_RANGE_USIZE.rev() {
-            if !self.indexes_by_resolution[resolution].is_empty() {
+            if !self.cells_by_resolution[resolution].is_empty() {
                 return Some(resolution as u8);
             }
         }
@@ -217,7 +194,7 @@ impl<'a> H3CompactedVec {
     /// compact all resolution from the given to 0
     ///
     /// resolutions are skipped when the compacting of the
-    /// former finer resolution added no new indexes to
+    /// former finer resolution added no new cells to
     /// the parent resolution unless include_resolutions
     /// forces the recompacting of a given resolution
     fn compact_from_resolution_up(&mut self, resolution: usize, include_resolutions: &[usize]) {
@@ -226,50 +203,48 @@ impl<'a> H3CompactedVec {
 
         for res in ((H3_MIN_RESOLUTION as usize)..=resolution).rev() {
             if !resolutions_touched.contains(&res) {
-                // no new indexes have been added to this resolution
+                // no new cells have been added to this resolution
                 continue;
             }
 
-            let mut indexes_to_compact = std::mem::take(&mut self.indexes_by_resolution[res]);
-            indexes_to_compact.sort_unstable();
-            indexes_to_compact.dedup();
-            let compacted = compact(&indexes_to_compact);
-            for h3_index in compacted {
-                let res = H3Cell::new(h3_index).resolution() as usize;
+            let mut cells_to_compact = std::mem::take(&mut self.cells_by_resolution[res]);
+            cells_to_compact.sort_unstable();
+            cells_to_compact.dedup();
+            let compacted = compact(&cells_to_compact);
+            for cell in compacted {
+                let res = cell.resolution() as usize;
                 resolutions_touched.insert(res);
-                self.indexes_by_resolution[res].push(h3_index);
+                self.cells_by_resolution[res].push(cell);
             }
         }
         self.purge_children();
     }
 
-    /// purge children of h3indexes already contained in lower resolutions
+    /// purge children of cells already contained in lower resolutions
     fn purge_children(&mut self) {
         let mut lowest_resolution = None;
-        for (r, h3indexes) in self.indexes_by_resolution.iter().enumerate() {
-            if lowest_resolution.is_none() && !h3indexes.is_empty() {
+        for (r, cells) in self.cells_by_resolution.iter().enumerate() {
+            if lowest_resolution.is_none() && !cells.is_empty() {
                 lowest_resolution = Some(r);
                 break;
             }
         }
 
         if let Some(lowest_res) = lowest_resolution {
-            let mut known_indexes = self.indexes_by_resolution[lowest_res]
+            let mut known_cells = self.cells_by_resolution[lowest_res]
                 .iter()
                 .cloned()
                 .collect::<HashSet<_>>();
 
             for r in (lowest_res + 1)..=(H3_MAX_RESOLUTION as usize) {
-                let mut orig_h3indexes = std::mem::take(&mut self.indexes_by_resolution[r]);
-                orig_h3indexes.drain(..).for_each(|h3index| {
-                    let index = H3Cell::new(h3index);
+                let mut orig_cells = std::mem::take(&mut self.cells_by_resolution[r]);
+                orig_cells.drain(..).for_each(|cell| {
                     let is_parent_known = (lowest_res..r).any(|parent_res| {
-                        known_indexes
-                            .contains(&index.get_parent_unchecked(parent_res as u8).h3index())
+                        known_cells.contains(&cell.get_parent_unchecked(parent_res as u8))
                     });
                     if !is_parent_known {
-                        known_indexes.insert(h3index);
-                        self.indexes_by_resolution[r].push(h3index);
+                        known_cells.insert(cell);
+                        self.cells_by_resolution[r].push(cell);
                     }
                 });
             }
@@ -277,54 +252,42 @@ impl<'a> H3CompactedVec {
     }
 }
 
-impl Default for H3CompactedVec {
+impl Default for CompactedCellVec {
     fn default() -> Self {
-        H3CompactedVec::new()
+        CompactedCellVec::new()
     }
 }
 
-impl FromIterator<H3Index> for H3CompactedVec {
-    fn from_iter<T: IntoIterator<Item = H3Index>>(iter: T) -> Self {
-        let mut cv = Self::new();
-        cv.add_indexes_from_iter(iter, true);
-        cv
-    }
-}
-
-impl FromIterator<H3Cell> for H3CompactedVec {
+impl FromIterator<H3Cell> for CompactedCellVec {
     fn from_iter<T: IntoIterator<Item = H3Cell>>(iter: T) -> Self {
         let mut cv = Self::new();
-        for index in iter {
-            cv.add_index(index.h3index(), false);
-        }
+        cv.add_cells(iter, false);
         cv.compact();
         cv
     }
 }
 
-impl From<Vec<H3Index>> for H3CompactedVec {
-    fn from(mut in_vec: Vec<H3Index>) -> Self {
+impl From<Vec<H3Cell>> for CompactedCellVec {
+    fn from(mut in_vec: Vec<H3Cell>) -> Self {
         let mut cv = Self::new();
-        for h3index in in_vec.drain(..) {
-            cv.add_index(h3index, false);
-        }
+        cv.add_cells(in_vec.drain(..), false);
         cv.compact();
         cv
     }
 }
 
-pub struct H3CompactedVecCompactedIterator<'a> {
-    compacted_vec: &'a H3CompactedVec,
+pub struct CompactedCellVecCompactedIterator<'a> {
+    compacted_vec: &'a CompactedCellVec,
     current_resolution: usize,
     current_pos: usize,
 }
 
-impl<'a> Iterator for H3CompactedVecCompactedIterator<'a> {
-    type Item = H3Index;
+impl<'a> Iterator for CompactedCellVecCompactedIterator<'a> {
+    type Item = H3Cell;
 
     fn next(&mut self) -> Option<Self::Item> {
         while self.current_resolution <= (H3_MAX_RESOLUTION as usize) {
-            if let Some(value) = self.compacted_vec.indexes_by_resolution[self.current_resolution]
+            if let Some(value) = self.compacted_vec.cells_by_resolution[self.current_resolution]
                 .get(self.current_pos)
             {
                 self.current_pos += 1;
@@ -338,35 +301,32 @@ impl<'a> Iterator for H3CompactedVecCompactedIterator<'a> {
     }
 }
 
-pub struct H3CompactedVecUncompactedIterator<'a> {
-    compacted_vec: &'a H3CompactedVec,
+pub struct CompactedCellVecUncompactedIterator<'a> {
+    compacted_vec: &'a CompactedCellVec,
     current_resolution: usize,
     current_pos: usize,
-    current_uncompacted: Vec<H3Index>,
+    current_uncompacted: Vec<H3Cell>,
     iteration_resolution: usize,
 }
 
-impl<'a> Iterator for H3CompactedVecUncompactedIterator<'a> {
-    type Item = H3Index;
+impl<'a> Iterator for CompactedCellVecUncompactedIterator<'a> {
+    type Item = H3Cell;
 
     fn next(&mut self) -> Option<Self::Item> {
         while self.current_resolution <= self.iteration_resolution {
             if self.current_resolution == self.iteration_resolution {
-                let value = self.compacted_vec.indexes_by_resolution[self.current_resolution]
+                let value = self.compacted_vec.cells_by_resolution[self.current_resolution]
                     .get(self.current_pos);
                 self.current_pos += 1;
                 return value.cloned();
             } else if let Some(next) = self.current_uncompacted.pop() {
                 return Some(next);
-            } else if let Some(next_parent) = self.compacted_vec.indexes_by_resolution
+            } else if let Some(next_parent) = self.compacted_vec.cells_by_resolution
                 [self.current_resolution]
                 .get(self.current_pos)
             {
-                self.current_uncompacted = H3Cell::new(*next_parent)
-                    .get_children(self.iteration_resolution as u8)
-                    .iter()
-                    .map(|i| i.h3index())
-                    .collect();
+                self.current_uncompacted =
+                    next_parent.get_children(self.iteration_resolution as u8);
                 self.current_pos += 1;
                 continue;
             } else {
@@ -380,27 +340,29 @@ impl<'a> Iterator for H3CompactedVecUncompactedIterator<'a> {
 
 #[cfg(test)]
 mod tests {
+    use std::convert::TryInto;
+
     use bincode::{deserialize, serialize};
 
-    use crate::collections::H3CompactedVec;
+    use crate::collections::CompactedCellVec;
 
     #[test]
     fn compactedvec_is_empty() {
-        let mut cv = H3CompactedVec::new();
+        let mut cv = CompactedCellVec::new();
         assert!(cv.is_empty());
         assert_eq!(cv.len(), 0);
-        cv.add_index(0x89283080ddbffff_u64, false);
+        cv.add_cell(0x89283080ddbffff_u64.try_into().unwrap(), false);
         assert!(!cv.is_empty());
         assert_eq!(cv.len(), 1);
     }
 
     #[test]
     fn compactedvec_serde_roundtrip() {
-        let mut cv = H3CompactedVec::new();
-        cv.add_index(0x89283080ddbffff_u64, false);
+        let mut cv = CompactedCellVec::new();
+        cv.add_cell(0x89283080ddbffff_u64.try_into().unwrap(), false);
         let serialized_data = serialize(&cv).unwrap();
 
-        let cv_2: H3CompactedVec = deserialize(&serialized_data).unwrap();
+        let cv_2: CompactedCellVec = deserialize(&serialized_data).unwrap();
         assert_eq!(cv, cv_2);
     }
 }