+++ h3ron/src/to_h3.rs
@@ -5,44 +5,43 @@ use geo::{
 use geo_types::{Coordinate, Geometry, Line, Polygon};
 
 use crate::error::check_valid_h3_resolution;
-use crate::{line, polyfill, Error, H3Cell, Index};
+use crate::{line, polyfill, Error, H3Cell};
 
 /// convert to indexes at the given resolution
 ///
 /// The output vec may contain duplicate indexes in case of
 /// overlapping input geometries.
-pub trait ToH3Indexes {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error>;
+pub trait ToH3Cells {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error>;
 }
 
-impl ToH3Indexes for Polygon<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for Polygon<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         check_valid_h3_resolution(h3_resolution)?;
-        let mut indexes = polyfill(self, h3_resolution);
-        Ok(indexes.drain(..).map(H3Cell::new).collect())
+        Ok(polyfill(self, h3_resolution))
     }
 }
 
-impl ToH3Indexes for MultiPolygon<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for MultiPolygon<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         let mut outvec = vec![];
         for poly in self.0.iter() {
-            let mut thisvec = poly.to_h3_indexes(h3_resolution)?;
+            let mut thisvec = poly.to_h3_cells(h3_resolution)?;
             outvec.append(&mut thisvec);
         }
         Ok(outvec)
     }
 }
 
-impl ToH3Indexes for Point<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for Point<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         check_valid_h3_resolution(h3_resolution)?;
         Ok(vec![H3Cell::from_coordinate(&self.0, h3_resolution)?])
     }
 }
 
-impl ToH3Indexes for MultiPoint<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for MultiPoint<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         let mut outvec = vec![];
         for pt in self.0.iter() {
             outvec.push(H3Cell::from_coordinate(&pt.0, h3_resolution)?);
@@ -51,74 +50,73 @@ impl ToH3Indexes for MultiPoint<f64> {
     }
 }
 
-impl ToH3Indexes for Coordinate<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for Coordinate<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         check_valid_h3_resolution(h3_resolution)?;
         Ok(vec![H3Cell::from_coordinate(self, h3_resolution)?])
     }
 }
 
-impl ToH3Indexes for LineString<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for LineString<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         check_valid_h3_resolution(h3_resolution)?;
-        let mut indexes = line(self, h3_resolution)?;
-        Ok(indexes.drain(..).map(H3Cell::new).collect())
+        line(self, h3_resolution)
     }
 }
 
-impl ToH3Indexes for MultiLineString<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for MultiLineString<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         let mut outvec = vec![];
         for ls in self.0.iter() {
-            let mut thisvec = ls.to_h3_indexes(h3_resolution)?;
+            let mut thisvec = ls.to_h3_cells(h3_resolution)?;
             outvec.append(&mut thisvec);
         }
         Ok(outvec)
     }
 }
 
-impl ToH3Indexes for Rect<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
-        self.to_polygon().to_h3_indexes(h3_resolution)
+impl ToH3Cells for Rect<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+        self.to_polygon().to_h3_cells(h3_resolution)
     }
 }
 
-impl ToH3Indexes for Triangle<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
-        self.to_polygon().to_h3_indexes(h3_resolution)
+impl ToH3Cells for Triangle<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+        self.to_polygon().to_h3_cells(h3_resolution)
     }
 }
 
-impl ToH3Indexes for Line<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
-        LineString::from(vec![self.start, self.end]).to_h3_indexes(h3_resolution)
+impl ToH3Cells for Line<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+        LineString::from(vec![self.start, self.end]).to_h3_cells(h3_resolution)
     }
 }
 
-impl ToH3Indexes for GeometryCollection<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for GeometryCollection<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         let mut outvec = vec![];
         for geom in self.0.iter() {
-            let mut thisvec = geom.to_h3_indexes(h3_resolution)?;
+            let mut thisvec = geom.to_h3_cells(h3_resolution)?;
             outvec.append(&mut thisvec);
         }
         Ok(outvec)
     }
 }
 
-impl ToH3Indexes for Geometry<f64> {
-    fn to_h3_indexes(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
+impl ToH3Cells for Geometry<f64> {
+    fn to_h3_cells(&self, h3_resolution: u8) -> Result<Vec<H3Cell>, Error> {
         match self {
-            Geometry::Point(pt) => pt.to_h3_indexes(h3_resolution),
-            Geometry::Line(l) => l.to_h3_indexes(h3_resolution),
-            Geometry::LineString(ls) => ls.to_h3_indexes(h3_resolution),
-            Geometry::Polygon(poly) => poly.to_h3_indexes(h3_resolution),
-            Geometry::MultiPoint(mp) => mp.to_h3_indexes(h3_resolution),
-            Geometry::MultiLineString(mls) => mls.to_h3_indexes(h3_resolution),
-            Geometry::MultiPolygon(mpoly) => mpoly.to_h3_indexes(h3_resolution),
-            Geometry::GeometryCollection(gc) => gc.to_h3_indexes(h3_resolution),
-            Geometry::Rect(r) => r.to_h3_indexes(h3_resolution),
-            Geometry::Triangle(tr) => tr.to_h3_indexes(h3_resolution),
+            Geometry::Point(pt) => pt.to_h3_cells(h3_resolution),
+            Geometry::Line(l) => l.to_h3_cells(h3_resolution),
+            Geometry::LineString(ls) => ls.to_h3_cells(h3_resolution),
+            Geometry::Polygon(poly) => poly.to_h3_cells(h3_resolution),
+            Geometry::MultiPoint(mp) => mp.to_h3_cells(h3_resolution),
+            Geometry::MultiLineString(mls) => mls.to_h3_cells(h3_resolution),
+            Geometry::MultiPolygon(mpoly) => mpoly.to_h3_cells(h3_resolution),
+            Geometry::GeometryCollection(gc) => gc.to_h3_cells(h3_resolution),
+            Geometry::Rect(r) => r.to_h3_cells(h3_resolution),
+            Geometry::Triangle(tr) => tr.to_h3_cells(h3_resolution),
         }
     }
 }