+++ h3ron/src/util.rs
@@ -3,15 +3,9 @@ use geo_types::{Coordinate, LineString, Point};
 use crate::Index;
 use h3ron_h3_sys::{degsToRads, GeoCoord, H3Index};
 
-/// filter out 0 values ( = positions in the vec not used to store h3indexes)
-macro_rules! remove_zero_indexes_from_vec {
-    ($h3_indexes:expr) => {
-        $h3_indexes.retain(|h3i: &H3Index| *h3i != 0);
-    };
-}
-
 #[inline]
 pub(crate) fn drain_h3indexes_to_indexes<T: Index>(mut v: Vec<H3Index>) -> Vec<T> {
+    // filters out 0 values ( = positions in the vec not used to store h3indexes, created by polyfill, k_ring, ...)
     v.drain(..).filter(|h3i| *h3i != 0).map(T::new).collect()
 }