+++ h3ronpy/src/collections.rs
@@ -1,14 +1,14 @@
 use pyo3::prelude::*;
 
+use crate::cells_to_h3indexes;
 use crate::error::IntoPyResult;
-use h3ron::collections as h3c;
 use h3ron::error::check_valid_h3_resolution;
-use h3ron_h3_sys::H3Index;
+use h3ron::{collections as h3c, H3Cell, Index};
 use numpy::{IntoPyArray, PyArray1};
 
 #[pyclass]
 pub struct H3CompactedVec {
-    pub(crate) inner: h3c::H3CompactedVec,
+    pub(crate) inner: h3c::CompactedCellVec,
 }
 
 #[pymethods]
@@ -30,32 +30,33 @@ impl H3CompactedVec {
     ///
     /// This function is pretty inefficient.
     fn contains(&self, h3index: u64) -> bool {
-        self.inner.contains(h3index)
+        self.inner.contains(H3Cell::new(h3index))
     }
 
     fn compacted_indexes(&self) -> Py<PyArray1<u64>> {
-        let indexes: Vec<_> = self.inner.iter_compacted_indexes().collect();
-        return_h3indexes_array(indexes)
+        let indexes: Vec<_> = self.inner.iter_compacted_cells().collect();
+        return_h3index_array(indexes)
     }
 
     fn compacted_indexes_at_resolution(&self, h3_resolution: u8) -> PyResult<Py<PyArray1<u64>>> {
         check_valid_h3_resolution(h3_resolution).into_pyresult()?;
-        let indexes = self
+        let cells = self
             .inner
-            .get_compacted_indexes_at_resolution(h3_resolution)
+            .get_compacted_cells_at_resolution(h3_resolution)
             .to_vec();
-        Ok(return_h3indexes_array(indexes))
+        Ok(return_h3index_array(cells))
     }
 
     fn uncompacted_indexes_at_resolution(&self, h3_resolution: u8) -> PyResult<Py<PyArray1<u64>>> {
         check_valid_h3_resolution(h3_resolution).into_pyresult()?;
-        let indexes = self.inner.iter_uncompacted_indexes(h3_resolution).collect();
-        Ok(return_h3indexes_array(indexes))
+        let cells = self.inner.iter_uncompacted_cells(h3_resolution).collect();
+        Ok(return_h3index_array(cells))
     }
 }
 
-fn return_h3indexes_array(h3indexes: Vec<H3Index>) -> Py<PyArray1<u64>> {
+#[inline]
+fn return_h3index_array(cells: Vec<H3Cell>) -> Py<PyArray1<u64>> {
     let gil = Python::acquire_gil();
     let py = gil.python();
-    h3indexes.into_pyarray(py).to_owned()
+    cells_to_h3indexes(cells).into_pyarray(py).to_owned()
 }