+++ h3ronpy/src/vector.rs
@@ -5,7 +5,7 @@ use pyo3::exceptions::PyValueError;
 use pyo3::{prelude::*, wrap_pyfunction};
 use rayon::prelude::*;
 
-use h3ron::{compact, Index, ToH3Indexes};
+use h3ron::{compact, Index, ToH3Cells};
 
 use crate::error::IntoPyResult;
 
@@ -13,18 +13,16 @@ fn wkbbytes_to_h3(wkbdata: &&[u8], h3_resolution: u8, do_compact: bool) -> PyRes
     let mut cursor = Cursor::new(wkbdata);
     match wkb::wkb_to_geom(&mut cursor) {
         Ok(g) => {
-            let mut indexes = g.to_h3_indexes(h3_resolution).into_pyresult()?;
-            let mut h3indexes: Vec<_> = indexes.drain(..).map(|i| i.h3index()).collect();
+            let mut cells = g.to_h3_cells(h3_resolution).into_pyresult()?;
 
             // deduplicate, in the case of overlaps or lines
-            h3indexes.sort_unstable();
-            h3indexes.dedup();
+            cells.sort_unstable();
+            cells.dedup();
 
-            if !do_compact {
-                Ok(h3indexes)
-            } else {
-                Ok(compact(&h3indexes))
+            if do_compact {
+                cells = compact(&cells);
             }
+            Ok(cells.drain(..).map(|i| i.h3index()).collect())
         }
         Err(err) => Err(PyValueError::new_err(format!("invalid WKB: {:?}", err))),
     }